+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.3 Francois Bessette
+
+- Add an unsubscribe link at the botton of emails sent when a new event
+  is published. Add unsubscribe form and store user preference to meta
+  variable called acc_email_on_new_events. Do not send email to users
+  who have variable set to 0.
+
 ## 1.2 Francois Bessette
 
 - The Brevo transactional email service accepts a maximum of 99 recipients.